requirements-depricated.txt
pytz
oslo.config>=3.14.0  # Apache-2.0

aodhclient>=0.5.0  # Apache-2.0
python-cinderclient>=1.6.0,!=1.7.0,!=1.7.1  # Apache-2.0
python-ceilometerclient>=2.5.0  # Apache-2.0
python-keystoneclient>=2.0.0,!=2.1.0  # Apache-2.0
python-muranoclient>=0.8.2  # Apache-2.0
python-novaclient>=2.29.0,!=2.33.0  # Apache-2.0
python-neutronclient>=5.1.0  # Apache-2.0
python-heatclient>=1.4.0  # Apache-2.0
python-glanceclient>=2.3.0,!=2.4.0  # Apache-2.0
python-saharaclient>=0.18.0  # Apache-2.0

Babel>=2.3.4  # BSD
paramiko>=2.0 # LGPLv2.1+
pbr>=1.6  # Apache-2.0
requests>=2.10.0  # Apache-2.0
unittest2  # BSD
PyYAML>=3.1.0  # MIT
testresources>=0.2.4  # Apache-2.0/BSD
nose  # LGPL
SQLAlchemy>=1.0.10,<1.1.0  # MIT
alembic>=0.8.4  # MIT
amqp>=1.4.0,<2.0 # LGPL
anyjson>=0.3.3  # BSD
gevent
keystonemiddleware>=4.0.0,!=4.1.0,!=4.5.0  # Apache-2.0
kombu>=3.0.25  # BSD
oslo.serialization>=1.10.0              # Apache-2.0
oslo.utils>=3.16.0                       # Apache-2.0
pecan>=1.0.0,!=1.0.2,!=1.0.3,!=1.0.4  # BSD
psycopg2>=2.5  # LGPL/ZPL
stevedore>=1.16.0  # Apache-2.0

positional
WebTest>=2.0.17
mock==1.0.1
nose-timer>=0.4.3
requests-mock>=0.5.1
tox>=1.7.1
coverage==3.6
pyparsing
monotonic
funcsigs
netifaces
ipdb
cryptography
pytest
